requirements.txt
streamlit
python-dotenv
pymongo
PyPDF2
langchain
langchain-google-genai
langchain-mongodb
langchain-text-splitters